tamil_news_aggregator/requirements.txt
asgiref==3.8.1
django==4.2.23
djangorestframework==3.15.2
greenlet==3.1.1
playwright==1.48.0
psycopg2-binary==2.9.10
pyee==12.0.0
python-decouple==3.8
pytz==2025.2
sqlparse==0.5.3
typing-extensions==4.13.2
